# web/src/lib/openapi.schema.json
{
  "components": {
    "schemas": {
...
      "CacheAddRequest": {
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "query"
        ],
        "title": "CacheAddRequest",
        "type": "object"
      },
      "CacheAddResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "query_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query Hash"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "supported": {
            "title": "Supported",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "supported",
          "query"
        ],
        "title": "CacheAddResponse",
        "type": "object"
      },
      "CacheDeleteResponse": {
        "properties": {
          "cache_id": {
            "title": "Cache Id",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "cache_id"
        ],
        "title": "CacheDeleteResponse",
        "type": "object"
      },
      "CacheDeployRequest": {
        "properties": {
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "mode": {
            "default": "docker",
            "title": "Mode",
            "type": "string"
          },
          "namespace": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Namespace"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "ssh_user": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssh User"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "title": "CacheDeployRequest",
        "type": "object"
      },
      "CacheDropAllResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "count"
        ],
        "title": "CacheDropAllResponse",
        "type": "object"
      },
      "CacheEntryResponse": {
        "properties": {
          "cache_id": {
            "title": "Cache Id",
            "type": "string"
          },
          "cache_name": {
            "title": "Cache Name",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "registry_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registry Hash"
          },
          "ttl": {
            "title": "Ttl",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "cache_id",
          "cache_name",
          "query",
          "type",
          "ttl"
        ],
        "title": "CacheEntryResponse",
        "type": "object"
      },
      "CacheErrorResponse": {
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "error"
        ],
        "title": "CacheErrorResponse",
        "type": "object"
      },
      "CacheListResponse": {
        "properties": {
          "caches": {
            "items": {
              "$ref": "#/components/schemas/CacheEntryResponse"
            },
            "title": "Caches",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "caches",
          "count"
        ],
        "title": "CacheListResponse",
        "type": "object"
      },
      "CacheRegisterRequest": {
        "properties": {
          "cache_host": {
            "title": "Cache Host",
            "type": "string"
          },
          "cache_port": {
            "default": 5433,
            "title": "Cache Port",
            "type": "integer"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "required": [
          "cache_host"
        ],
        "title": "CacheRegisterRequest",
        "type": "object"
      },
...
      "CacheRunRequest": {
        "properties": {
          "iterations": {
            "default": 15,
            "title": "Iterations",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "warmup": {
            "default": 5,
            "title": "Warmup",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "CacheRunRequest",
        "type": "object"
      },
      "CacheStatusResponse": {
        "properties": {
          "cache_target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Target"
          },
          "container_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Container Name"
          },
          "deployed": {
            "title": "Deployed",
            "type": "boolean"
          },
          "endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint"
          },
          "running": {
            "title": "Running",
            "type": "boolean"
          }
        },
        "required": [
          "deployed",
          "running"
        ],
        "title": "CacheStatusResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "API server for RDST web client",
    "title": "RDST API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cache/add": {
      "post": {
        "description": "Create a cache or dry-run check.",
        "operationId": "add_cache_api_cache_add_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CacheAddResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CacheErrorResponse"
                    }
                  ],
                  "title": "Response Add Cache Api Cache Add Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Cache",
        "tags": [
          "cache",
          "cache"
        ]
      }
    },
    "/api/cache/deploy": {
      "post": {
        "description": "Deploy ReadySet cache (SSE stream).",
        "operationId": "deploy_cache_api_cache_deploy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheDeployRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deploy Cache",
        "tags": [
          "cache",
          "cache"
        ]
      }
    },
    "/api/cache/drop-all": {
      "delete": {
        "description": "Delete all caches.",
        "operationId": "drop_all_caches_api_cache_drop_all_delete",
        "parameters": [
          {
            "description": "Target database name",
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "description": "Target database name",
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CacheDropAllResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CacheErrorResponse"
                    }
                  ],
                  "title": "Response Drop All Caches Api Cache Drop All Delete"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Drop All Caches",
        "tags": [
          "cache",
          "cache"
        ]
      }
    },
    "/api/cache/list": {
      "get": {
        "description": "List cached queries for a database target.",
        "operationId": "list_caches_api_cache_list_get",
        "parameters": [
          {
            "description": "Target database name",
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "description": "Target database name",
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CacheListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CacheErrorResponse"
                    }
                  ],
                  "title": "Response List Caches Api Cache List Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Caches",
        "tags": [
          "cache",
          "cache"
        ]
      }
    },
    "/api/cache/register": {
      "post": {
        "description": "Register a cache target with a user-provided endpoint (for non-local deploys).",
        "operationId": "register_cache_target_api_cache_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CacheStatusResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CacheErrorResponse"
                    }
                  ],
                  "title": "Response Register Cache Target Api Cache Register Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Cache Target",
        "tags": [
          "cache",
          "cache"
        ]
      }
    },
    "/api/cache/remove": {
      "delete": {
        "description": "Remove cache target and stop container.",
        "operationId": "remove_cache_api_cache_remove_delete",
        "parameters": [
          {
            "description": "Target database name",
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "description": "Target database name",
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CacheDeleteResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CacheErrorResponse"
                    }
                  ],
                  "title": "Response Remove Cache Api Cache Remove Delete"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Cache",
        "tags": [
          "cache",
          "cache"
        ]
      }
    },
    "/api/cache/run": {
      "post": {
        "description": "Run a query against both origin DB and ReadySet cache, stream comparison results (SSE).",
        "operationId": "run_cache_comparison_api_cache_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Cache Comparison",
        "tags": [
          "cache",
          "cache"
        ]
      }
    },
    "/api/cache/status": {
      "get": {
        "description": "Check cache deployment status for a database target.",
        "operationId": "get_cache_status_api_cache_status_get",
        "parameters": [
          {
            "description": "Target database name",
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "description": "Target database name",
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CacheStatusResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CacheErrorResponse"
                    }
                  ],
                  "title": "Response Get Cache Status Api Cache Status Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Cache Status",
        "tags": [
          "cache",
          "cache"
        ]
      }
    },
    "/api/cache/{cache_id}": {
      "delete": {
        "description": "Delete a single cache by ID.",
        "operationId": "delete_cache_api_cache__cache_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "cache_id",
            "required": true,
            "schema": {
              "title": "Cache Id",
              "type": "string"
            }
          },
          {
            "description": "Target database name",
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "description": "Target database name",
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/CacheDeleteResponse"
                    },
                    {
                      "$ref": "#/components/schemas/CacheErrorResponse"
                    }
                  ],
                  "title": "Response Delete Cache Api Cache  Cache Id  Delete"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Cache",
        "tags": [
          "cache",
          "cache"
        ]
      }
    },
...
  }
}